Fix a sign-compare warning

+++ devicecache.cpp
@@ -134,9 +134,9 @@ void DeviceCache::checkDevice ( Solid::Device solidDevice )
                 {
                     LIBMTP_raw_device_t* rawDevice = &rawdevices[i];
                     uint32_t rawBusNum = rawDevice->bus_location;
-                    uint32_t rawDevNum = rawDevice->devnum;
+                    uint8_t rawDevNum = rawDevice->devnum;
 
-                    if ( rawBusNum == solidBusNum && rawDevNum == solidDevNum )
+                    if ( rawBusNum == solidBusNum && static_cast<int>(rawDevNum) == solidDevNum )
                     {
                         kDebug( KIO_MTP ) << "Found device matching the Solid description";
